--- backend/framework/alembic/versions/76a287dd6f88_edit_kevinwalshmunozmonsters_fields.py
@@ -1,38 +1,0 @@
-"""edit kevinWalshMunozMonsters fields
-
-Revision ID: 76a287dd6f88
-Revises: b872c2842a95
-Create Date: 2025-08-24 23:34:32.012017
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '76a287dd6f88'
-down_revision: Union[str, Sequence[str], None] = 'b872c2842a95'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('kevinWalshMunozMonsters', sa.Column('index', sa.String(), nullable=False))
-    op.add_column('kevinWalshMunozMonsters', sa.Column('url', sa.String(), nullable=False))
-    op.drop_column('kevinWalshMunozMonsters', 'id')
-    op.drop_column('kevinWalshMunozMonsters', 'description')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('kevinWalshMunozMonsters', sa.Column('description', sa.VARCHAR(), autoincrement=False, nullable=True))
-    op.add_column('kevinWalshMunozMonsters', sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False))
-    op.drop_column('kevinWalshMunozMonsters', 'url')
-    op.drop_column('kevinWalshMunozMonsters', 'index')
-    # ### end Alembic commands ###
