--- backend/framework/alembic/versions/b872c2842a95_add_kevinwalshmunozmonsters_table.py
@@ -1,37 +1,0 @@
-"""add kevinWalshMunozMonsters table
-
-Revision ID: b872c2842a95
-Revises: aa75144a150e
-Create Date: 2025-08-24 23:30:22.137278
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'b872c2842a95'
-down_revision: Union[str, Sequence[str], None] = 'aa75144a150e'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('kevinWalshMunozMonsters',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('name', sa.String(), nullable=False),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('kevinWalshMunozMonsters')
-    # ### end Alembic commands ###
